moved the handleSearch function up to the home.jsx module and passed the function down from home.jsx to header.jsx and then to searchForm.jsx and passed in the form inputs up to the handleSearch method up to home.jsx using props
