+++ client/components/searchResults.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+export default class SearchResults extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return null;
+  }
+}
